+++ package.json
@@ -136,7 +136,7 @@
         "uuid": "11.1.0",
         "winston": "3.17.0",
         "xxhash-wasm": "1.1.0",
-        "youtubei.js": "^13.4.0",
+        "youtubei.js": "13.4.0",
         "zod": "3.24.4"
     },
     "devDependencies": {
